+++ plugin-maven/src/main/java/com/diffplug/spotless/maven/FormatterFactory.java
@@ -85,13 +85,12 @@ public abstract class FormatterFactory {
 				.collect(Collectors.toCollection(() -> new ArrayList<FormatterStep>()));
 		if (toggle != null) {
 			List<FormatterStep> formatterStepsBeforeToggle = formatterSteps;
-			formatterSteps = Collections.singletonList(toggle.createFence().preserveWithin(config.getFileLocator().getBaseDir().toPath(), formatterStepsBeforeToggle));
+			formatterSteps = Collections.singletonList(toggle.createFence().preserveWithin(formatterStepsBeforeToggle));
 		}
 		return Formatter.builder()
 				.encoding(formatterEncoding)
 				.lineEndingsPolicy(formatterLineEndingPolicy)
 				.steps(formatterSteps)
-				.rootDir(config.getFileLocator().getBaseDir().toPath())
 				.build();
 	}