+++ testlib/src/test/java/com/diffplug/spotless/FormatterTest.java
@@ -59,7 +59,6 @@ class FormatterTest {
 				return Formatter.builder()
 						.lineEndingsPolicy(lineEndingsPolicy)
 						.encoding(encoding)
-						.rootDir(rootDir)
 						.steps(steps)
 						.build();
 			}